Flight Paths: Apply reputation discount to all points in multi-segment paths

+++ src/game/MotionGenerators/WaypointMovementGenerator.h
@@ -135,7 +135,7 @@ class FlightPathMovementGenerator
         struct TaxiNodeChangeInfo
         {
             uint32 PathIndex;
-            int32 Cost;
+            int64  Cost;
         };
 
         std::deque<TaxiNodeChangeInfo> _pointsForPathSwitch;    //! node indexes and costs where TaxiPath changes
